+++ include/dirc_point.h
@@ -3,13 +3,13 @@
 #define DIRC_POINT
 struct dirc_point
 {
-	double x;
-	double y;
-	double t;
+	float x;
+	float y;
+	float t;
 	int updown;
 	int last_wall_x;
 	int wedge_before_interface;
-	double weight;
-	double init_phi; //internal validation only
+	float weight;
+	float init_phi; //internal validation only
 };
 #endif
