+++ include/dirc_probability_spread.h
@@ -9,7 +9,7 @@ class DircProbabilitySpread
 protected:
 	std::vector<dirc_point> support_points;
 	bool test_time_dir;
-	double get_weight(dirc_point inpoint);
+	float get_weight(dirc_point inpoint);
 
 	
 public:
@@ -18,8 +18,8 @@ public:
 		std::vector<dirc_point> isupport,\
 		bool itest_time_dir = true);
 	
-	inline double get_log_likelihood(std::vector<dirc_point> inpoints) __attribute__((always_inline));
-	double get_single_log_likelihood(dirc_point inpoint);
-	virtual double support_spread_function(dirc_point support, dirc_point test) = 0;
+	inline float get_log_likelihood(std::vector<dirc_point> inpoints) __attribute__((always_inline));
+	float get_single_log_likelihood(dirc_point inpoint);
+	virtual float support_spread_function(dirc_point support, dirc_point test) = 0;
 };
 #endif
